+++ src/insert/line_type.rs
@@ -1,20 +1,22 @@
 #[derive(Clone, Debug, PartialEq)]
 pub enum LineType {
+	Cancel,
+	Pick,
 	Exec,
 	Label,
-	Reset,
 	Merge,
-	Cancel,
+	Reset,
 }
 
 impl ToString for LineType {
 	fn to_string(&self) -> String {
 		match *self {
+			Self::Cancel => String::from("<cancel>"),
+			Self::Pick => String::from("pick"),
 			Self::Exec => String::from("exec"),
 			Self::Label => String::from("label"),
-			Self::Reset => String::from("reset"),
 			Self::Merge => String::from("merge"),
-			Self::Cancel => String::from("<cancel>"),
+			Self::Reset => String::from("reset"),
 		}
 	}
 }
@@ -28,11 +30,12 @@ mod tests {
 	#[rstest(
 		line_type,
 		expected,
+		case::cancel(&LineType::Cancel, "<cancel>"),
+		case::pick(&LineType::Pick, "pick"),
 		case::exec(&LineType::Exec, "exec"),
-		case::exec(&LineType::Label, "label"),
-		case::exec(&LineType::Reset, "reset"),
-		case::exec(&LineType::Merge, "merge"),
-		case::exec(&LineType::Cancel, "<cancel>"),
+		case::label(&LineType::Label, "label"),
+		case::merge(&LineType::Merge, "merge"),
+		case::reset(&LineType::Reset, "reset"),
 	)]
 	fn to_string(line_type: &LineType, expected: &str) {
 		assert_eq!(line_type.to_string(), String::from(expected));